many compile fix
git-svn-id: http://ltt.polymtl.ca/svn@851 04897980-b3bd-0310-b5e0-8ef037075253

+++ ltt/branches/poly/ltt/ltt.h
@@ -87,7 +87,7 @@ typedef struct _LttSystemDescription LttSystemDescription;
 /* Checksums are used to differentiate facilities which have the same name
    but differ. */
 
-typedef unsigned long LttChecksum;
+typedef guint32 LttChecksum;
 
 
 /* Events are usually stored with the easily obtained CPU clock cycle count,
